(PDB-3611) Deduplicate facts (again)
Previously, we had changed fact storage to put the value in the facts table, in
order to simplify storage and sidestep the problematic garbage collection issues
when fact values had their own table. But this proved to make queries too slow,
especially when dealing with queries which have multiple conditions on facts.

This commit contains the schema migration to move facts back to the deduplicated
table structure, with a separate fact_values table.